Add missing type aliases
