+++ app/routes/__init__.py
@@ -1,0 +1,5 @@
+from .auth import auth_bp
+from .dashboard import dashboard_bp
+from .shopping import shopping_bp
+
+__all__ = ['auth_bp', 'dashboard_bp', 'shopping_bp']
