+++ app/routes/dashboard.py
@@ -1,0 +1,9 @@
+from flask import Blueprint, render_template
+from flask_login import login_required
+
+dashboard_bp = Blueprint('dashboard', __name__)
+
+@dashboard_bp.route('/')
+@login_required
+def index():
+    return render_template('dashboard.html')
