+++ templates/accept_invitation.html
@@ -58,7 +58,7 @@
                             
                             <div class="d-grid gap-2">
                                 <button type="submit" class="btn btn-primary">Accept Invitation</button>
-                                <a href="{{ url_for('login') }}" class="btn btn-secondary">Cancel</a>
+                                <a href="{{ url_for('auth.login') }}" class="btn btn-secondary">Cancel</a>
                             </div>
                         </form>
                     {% else %}
@@ -69,7 +69,7 @@
                             <p class="mb-0">Please contact the account administrator for a new invitation.</p>
                         </div>
                         <div class="d-grid">
-                            <a href="{{ url_for('login') }}" class="btn btn-primary">Return to Login</a>
+                            <a href="{{ url_for('auth.login') }}" class="btn btn-primary">Return to Login</a>
                         </div>
                     {% endif %}
                 </div>
