+++ templates/generate_meal_plan.html
@@ -27,7 +27,7 @@
                         
                         <div class="d-grid gap-2">
                             <button type="submit" class="btn btn-primary">Generate Plan</button>
-                            <a href="{{ url_for('dashboard') }}" class="btn btn-secondary">Cancel</a>
+                            <a href="{{ url_for('dashboard.index') }}" class="btn btn-secondary">Cancel</a>
                         </div>
                     </form>
                 </div>
