+++ templates/login.html
@@ -40,7 +40,7 @@
                         </div>
                     </form>
                     <div class="text-center mt-3">
-                        <p>Don't have an account? <a href="{{ url_for('register') }}">Register here</a></p>
+                        <p>Don't have an account? <a href="{{ url_for('auth.register') }}">Register here</a></p>
                     </div>
                 </div>
             </div>
