+++ templates/view_recipe.html
@@ -8,7 +8,7 @@
         <span>Serves: {{ recipe.servings }}</span>
         {% if recipe.source_link %}| <a href="{{ recipe.source_link }}" target="_blank">Source Link</a>{% endif %}
          | <a href="{{ url_for('edit_recipe', recipe_id=recipe.id) }}">Edit Recipe</a>
-         | <a href="{{ url_for('dashboard') }}">Back to Meal Plan</a>
+         | <a href="{{ url_for('dashboard.index') }}">Back to Meal Plan</a>
     </div>
     <div class="recipe-details">
         <div class="recipe-ingredients">
